refactor: Car 의 name 과 distance 속성을 포장하도록 수정
- '모든 원시값과 문자열을 포장한다'는 객체지향 생활 체조 원칙에 따라 수정
