+++ src/main/java/racingcar/domain/Distance.java
@@ -1,0 +1,18 @@
+package racingcar.domain;
+
+public class Distance {
+
+    private int distance;
+
+    public Distance(int distance) {
+        this.distance = distance;
+    }
+
+    public void increase() {
+        distance++;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
